massive performance change
usually the program on a very small data set would not use any more than 2% cpu. in my case, i was testing 500 of them. due to the fact that a new instance was being created for each link being tested, cpu usage went as high as 50-60% used cpu when it achieved around the 300 mark. it even slowed down the performance of the program itself, to the point where it took approximately 7 HOURS to test all of the links. 

however, there was a very easy fix to this. a new vlc instance does NOT have to be initialized when testing a new link. you can simply define and set the media on the same instance. this drastically reduced the cpu usage back down to that <2% usage, and drastically improved the testing speed and it finished in just over an hour.